requirements.txt
torch
torchvision
torchaudio
pillow
numpy<2.00
matplotlib
tqdm
huggingface-hub
opencv-python
timm>=0.6.12
pandas==2.2.2
scikit_learn==1.5.0
flask
flask_restful
flask_cors